agent-rl-core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-rl-core"
version = "0.1.0"
description = "Shared infrastructure for verifier-based agent RL projects"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pyyaml>=6.0",
  "pydantic>=2.6",
  "numpy>=1.26",
  "tqdm>=4.66"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
